refactor: move react component files to /components
