+++ src/web/src/views/cli/components/CLIModGeneratorProfileTabs.tsx
