+++ src/web/src/views/cli/utils/commandTreeInitialization.ts
@@ -10,7 +10,7 @@ import {
   CLISpecsSimpleCommand,
   CLISpecsSimpleCommandGroup,
   CLISpecsSimpleCommandTree,
-} from "../CLIModuleGenerator";
+} from "../components/CLIModuleGenerator";
 import {
   calculateSelected,
   type ProfileCTCommandGroup,
